Add inline logging option
By default logs are asynchronously written to stderr. That is to say,
it's done off thread. This patch introduces a compile time build option
that forces logs to print inline. This is useful for unit tests, where
we do not want logs to be spewing output after the test has ended.