Add inline logging option (#21)
Summary:
By default logs are asynchronously written to stderr. That is to say,
it's done off thread. This patch introduces a compile time build option
that forces logs to print inline. This is useful for unit tests, where
we do not want logs to be spewing output after the test has ended.

Note this patch is currently branched off of another PR. I'll rebase
once the first one lands to clean it up.
Pull Request resolved: https://github.com/facebookincubator/oomd/pull/21

Differential Revision: D9845464

Pulled By: danobi

fbshipit-source-id: 7057d18a89aebb4b1689f93f11d8543e928d9704