Change core kill plugins to kill via kill list
Summary:
This changes the core kill plugins to only kill cgroups
in an explicit list. In other words, only the cgroups passed to
the `cgroup` argument will be considered for killing.

It's possible (and encouraged) to replicate the old kill behavior
by adding a `/*` to the end of the parent cgroups. See the changed
tests for examples.

Reviewed By: mokomull

Differential Revision: D13144192

fbshipit-source-id: 95665c999c5115d5c66cbf023cad44e363279475