Add wildcard support to detector plugins
Summary:
This patch adds wildcarded cgroup path support to the core
detector plugins. This is useful for monitoring cgroups where the
full name may not be known at oomd configuration time.

Note we only keep track of the highest memory pressure cgroup if
the wildcarded path globs to more than one cgroup. The reason behind
this is two fold:

    * We expect the common case of wildcard usage to only glob to
      one cgroup. If globbing to multiple cgroups is desired, it would
      make more sense to construct the wildcard paths such that every
      glob only maps to one cgroup.

    * In the event that the wildcard path globs to more than one
      cgroup, keeping track of the max([pressure(cgroup) for cgroup
      in globbed]) gives equal representation to each cgroup.

Reviewed By: htejun

Differential Revision: D13087317

fbshipit-source-id: 02930bbd5e525f2a81ead79df08e1adc72b0edc2