Reduce sleeps in core plugin tests
There used to be a lot of sleeps in the unit test code path.
This was because:
* the mocked out tryToKillPids would always "kill N pids", causing
  tryToKillCgroup to retry a bunch of times
* tryToKillCgroup was sleeping after each retry. This was to give
  the system time to respond to the SIGKILLs

The first issue is fixed in this patch using a set of "already killed"
pids. The second issue can be elided by removing the sleep after the
first set of kills (which isn't really necessary anyways).