Reduce sleeps in core plugin tests (#31)
Summary:
There used to be a lot of sleeps in the unit test code path.
This was because:
* the mocked out tryToKillPids would always "kill N pids", causing
  tryToKillCgroup to retry a bunch of times
* tryToKillCgroup was sleeping after each retry. This was to give
  the system time to respond to the SIGKILLs

The first issue is fixed in this patch using a set of "already killed"
pids. The second issue can be elided by removing the sleep after the
first set of kills (which isn't really necessary anyways).
Pull Request resolved: https://github.com/facebookincubator/oomd/pull/31

Differential Revision: D13366444

Pulled By: danobi

fbshipit-source-id: e702e07d971e7e6d3337d56f073732e8909ceffe