Add meson build and docs for memory_above and systemd_restart plugins
This patch adds open source build support for memory_above and
systemd_restart. systemd_restart is moved into its own directory because
it relies on dbus + systemd to function, and thus should not be
considered a core plugin.

The build is modified to only build systemd_restart and associated tests
if libsystemd is installed on the host.