Add meson build and docs for memory_above and systemd_restart plugins (#44)
Summary:
This patch adds open source build support for memory_above and
systemd_restart. systemd_restart is moved into its own directory because
it relies on dbus + systemd to function, and thus should not be
considered a core plugin.

The build is modified to only build systemd_restart and associated tests
if libsystemd is installed on the host.
Pull Request resolved: https://github.com/facebookincubator/oomd/pull/44

Reviewed By: abulimov

Differential Revision: D13521939

Pulled By: danobi

fbshipit-source-id: 9245d7b7dadc6f2b32693c9b5c4ccec268f3a0c7