Allow oomd to fail gracefully if PSI is not supported by kernel
Summary:
Before, when oomd is run and PSI is not supported on the kernel it causes a sigfault which causes systemd to try to restart it again and again.

The changes in this diff will have oomd first check for the presence of /proc/mempressure or /proc/pressure/memory. If neither exist, it will return an informative error message and exit with a status code of 3.

Reviewed By: mokomull

Differential Revision: D13578039

fbshipit-source-id: ae39761ccb17b7251d5363706daafaad883b2dc1