Return cleanly on errors instead of aborting (#43)
Summary:
Aborting is bad because it creates a core dump. In isolated cases, this
is not too bad. However, when thousands of hosts are crash looping, it
burns through a lot of disk. Instead, we should return cleanly and let
systemd handle failures.

There is also another important benefit of this patch: it forces the
async logging thread to shut down cleanly. With this, we can ensure
logs do not get stranded in the buffer at shutdown.
Pull Request resolved: https://github.com/facebookincubator/oomd/pull/43

Reviewed By: mokomull

Differential Revision: D13518604

Pulled By: danobi

fbshipit-source-id: 65b07781491d5167020a4f4f816d7d0986d22580