Do an unconditional size kill if nothing grows fast enough
Summary:
Before, if kill_by_memory_size_or_growth was triggered and a size
kill was failed (ie a cgroup did not use >50% of monitored parent cgroup),
something would be growth killed.

That doesn't make much sense if nothing is growing fast. This is actually
the common case where the system is just oversubscribed. This patch makes
it so kill_by_memory_size_or_growth falls back to killing whatever the
biggest cgroup is.

Reviewed By: htejun

Differential Revision: D13794197

fbshipit-source-id: 3566cf16b9e9e576f032816804ccea97f0f45c4d