Back event loop by epoll
Summary:
Previously, oomd used a while/sleep event loop. This worked fine, but
a timerfd is more precise. Also with timerfd we don't have to worry about
any drift within the event loop. The kernel will do the timekeeping for us.

This also lays the ground for drop in configs. Drop in configs will use
inotify so we don't have to poll the drop in directory.

Reviewed By: mokomull

Differential Revision: D14155264

fbshipit-source-id: a1c2c8b6831c91a6b06d8d01e72212a8aa2a3c48