Remove config versioning
Summary:
Versioning configs is kind of backwards to the normal way of
doing things. People usually expect the code to be versioned, and configs
to target major release versions of code. Versioning the configs is sort
of like versioned symbols in that both parties (code and config) are
screwed in really tightly to each other.

Differential Revision: D14761961

fbshipit-source-id: f6e87aaec1473345c6f907fed0466079613ea359