Fix swap kills during swapoff
Summary:
oomd was making swap kills during `swapoff -a`. This happened
because it takes the kernel a bit of time to page everything back into
memory from the disabled swap. During that period of time, swap_free
is set to 0 and swap_total slowly decreases. This makes oomd think swap
is at 100% usage.

This patch makes oomd a bit smarter about whether swap is turned on.

Differential Revision: D15006630

fbshipit-source-id: f40aa6dec0ec0fafe6562eb3c4c18dac65905c72