Fix travis error on FsTest
Travis seems to run tests in docker with a FS such that files and
directories are created in different mount points, so that the existing
ntfw function does not walk the files. There does not seem to have
obvious concern of removing temp files across mount points, so remove
the FTW_MOUNT flag to resolve the test failure.