Fix travis error and warnings
Summary:
Travis failed after D16921059 https://travis-ci.com/facebookincubator/oomd/builds/123959581

I made a PR on github so travis and verify that the failure and warnings are actually fixed: https://github.com/facebookincubator/oomd/pull/86

From one commit of the PR:
```
Travis seems to run tests in docker with a FS such that files and
directories are created in different mount points, so that the existing
ntfw function does not walk the files. There does not seem to have
obvious concern of removing temp files across mount points, so remove
the FTW_MOUNT flag to resolve the test failure.
```

This diff is exactly the same as the PR above. Not sure what is the best way to get this change properly tested and committed.

Differential Revision: D16933962

fbshipit-source-id: d5ca74fc04c115b6bdcf6e0890298f3b169780c5