Move removeSiblingCgroups into OomdContext
Summary:
It turns out this method was useful for more than just kill
plugins. So move it into OomdContext so all plugins can use it.

This patch updates all the call sites and also de-BaseKillPlugin's
MemoryReclaim. MemoryReclaim was only ever converted to a kill plugin
so it could use the method.

Reviewed By: dschatzberg

Differential Revision: D17066610

fbshipit-source-id: 9ce226582c7203cd2025b5727af0bddf13359e3d