Check for "hybrid" hierarchy in OomdTest::MonitorRootHost.
Summary:
The test only works when the cgroup tree is mounted using the "unified" hierarchy, but the test was simply checking for whether there was a mount for cgroup2 and this was breaking the test on Ubuntu Bionic outside of a Docker container.

Technically, it would be possible for this to work in the "hybrid" hierarchy if the memory cgroup was mounted under `unified/` (cgroup2 tree) rather than `memory/` (cgroup1 controller), but that's a very unusual setup and we're not really aiming to support hybrid anyways.

The check is a bit crude in that it's checking for a specific path, but that should be sufficient for now, especially since it's only a test case.

In the future, we might want to consider replacing the current check that parses `/proc/mounts` looking for a `cgroup2` mount with the procedure outlined here:
https://systemd.io/CGROUP_DELEGATION.html#three-different-tree-setups-

Reviewed By: danobi

Differential Revision: D17228913

fbshipit-source-id: 5ff3602c0d11a97679d7a1e8b608795ca349200e