Update Travis-CI configuration to use native support for Ubuntu Bionic
Summary:
Replace the configuration using a Docker container with one that uses Ubuntu Bionic directly (Bionic became available on Travis-CI quite recently.)

This should make our Travis-CI builds quite a bit faster and it should also fix some issues, such as the fact that the `clang` configuration was in effect using gcc since the build itself was happening inside Docker and the compiler variables were not propagated there.

Additionally, split out the `clang-format` step into its own rule and fix the `clang-format` script to actually show the diff when there is a problem with formatting that might need a local run of `clang-format` to fix.

Remove the `build-from-github.sh` script, which is no longer used or needed.

Reviewed By: danobi

Differential Revision: D17228912

fbshipit-source-id: 7e0a71e008a0f8bf528528c283d1cd6aee583246