Rename binary and ship config files.
Summary:
Let's ship it with an `oomd` binary rather than `oomd_bin`. Now that we moved the tree to `src/`, we are free to use that name in Meson.

Generate an `oomd.service` file from a template, so we can adjust paths depending on configuration settings.

Adjust it to use `/etc/oomd/oomd.json` by default as configuration file.

Ship the existing `desktop.json` as the default `oomd.json` in that directory.

Reviewed By: danobi

Differential Revision: D17241829

fbshipit-source-id: f8ab1f53433c80b6e2ef933cf4ad97cd730f5acd