Skip using Meson's `copy` which was only introduced in Meson 0.46.0
Summary:
Argument `copy : true` was only introduced recently, Ubuntu Bionic still ships Meson 0.45.0 and fails here.

Let's stick to using `configuration` for now, since that prevents it from failing. That should be fine, because the shipped JSON file will not match any of the patterns, so no substitutions will be made.

We could use an empty `configuration_data()`, but let's use the same one with the directory locations, since it will work just fine and we'll need to change fewer lines of build rules for this change.

Reviewed By: dschatzberg

Differential Revision: D17269430

fbshipit-source-id: 37f5b6b6869c8d28cab75efe5f9702faf9aa1cab