Fix warning cannot open io.stat (#106)
Summary:
Pull Request resolved: https://github.com/facebookincubator/oomd/pull/106

Some cgroups does not have io.stat file because io controller is not enabled. The existing way calls readIostat and when there is an exception, we catch and warn once.

However, this never works as expected because internally readFileByLine does not throw exception but instead log to OLOG. This causes some log pollution when some monitored cgroup has no io controller.

The right way is to check for io controller instead of catching exception.

Reviewed By: danobi

Differential Revision: D17634818

fbshipit-source-id: b07d5335cd0bed3c63e5201f6b432c3720031709