oomd: fs: Reduce CPU usage by up to 10% by avoiding lstat check
Summary:
On sample prod machines with moderately sized hierarchies, this saves about
~10% of oomd's CPU time. This also saves 7% of oomd's CPU time in my profiles
on my T470s with a small hierarchy, so the saving scales to the number of
cgroups.

We currently run lstat() on every single dirent that we come across inside the
cgroup hierarchy. This is fairly inefficient, since it requires repeatedly
transitioning to kernel space and going through the VFS and kernfs layers to do
a number of potentially costly checks.

In modern kernels, kernfs is kind enough to populate the d_type field with
something useful (ie. not DT_UNKNOWN). Since we already have the dirent with
d_type attached, with this patch we don't actually have to call lstat() at all
in the nominal case, which saves quite a bit of work and CPU time.

I've left the old code intact in case there are limited situations in which
d_type is unset, but this avoids a non-trivial amount of overhead in the vast
majority of cases.

Reviewed By: danobi, hnaz

Differential Revision: D18708936

fbshipit-source-id: 047d34aa34d535975571340bb2b8ebb7dab65ba1