oomd: Use ~1-1.5% less CPU by reworking EntryType to a bitflag
Summary:
readDir is one of the hottest functions in profiles. We generally call it
perfectly legitimately (although in one case in this diff I removed an unused
fetch for `EntryType::REG_FILE`), but we often call it multiple times when we
could do so once in order to differentiate files from directories, since we
pass an EntryType.

This nice compartmentalisation would be fine on a rarely called function, but
on something so hot, we should be willing to tradeoff a bit more towards
reducing CPU time. This patch changes the code so we return a new type, `struct
DirEnt`, and choose what to populate it with using `DirEntFlags`. This means we
can avoid iterating twice in kernfs, which reduces the amount of time both in
userspace and in kernelspace doing inode/permission/other work. All in all,
this saves about 1% to 1.5% in my tests on a system with ~150 cgroups.

Reviewed By: danobi

Differential Revision: D18720763

fbshipit-source-id: 3353b89c4e8ffe658cae78196f9055a6cd31e3e7