oomd: Validate interval before starting execution
Summary:
At one point during profiling oomd, I tried to run with `--interval 0.1` to try
and stress the cgroup hierarchy further. oomd succeeded and ostensibly started
running, but actually ended up running with an interval of 0, since `std::stoi`
just took the first part of the string and discarded the rest.

This patch makes sure that the following cases are handled and produce an
error:

1. Non-numeric
2. Non-integer
3. Integer < 0 (which blocks the event loop forever, or makes timerfd sad)

Reviewed By: dschatzberg

Differential Revision: D18740260

fbshipit-source-id: e453edec6aedbc596b67245e67d7ccddc6c16cfc