oomd: Take a CgroupPath& in Fs::resolveWildcardPath
Summary:
In every non-test caller, we have a CgroupPath available and just call
CgroupPath::absolutePath() before passing it in. This seems like it was
primarily to make tests simpler, but changing it to use a CgroupPath& directly
means we can both make the tests more closely match reality, and we can always
know that we will have a relative path.

This allows removing the relative path handling cod, and we can always know
that we will have a relative path. This allows removing the relative path
handling code and do less work.

Reviewed By: dschatzberg

Differential Revision: D18785020

fbshipit-source-id: 1e695038bf35e075fbe66e53a9f584436fd87ae7