oomd: Save ~2% CPU by short-circuiting kernfs read{,dir}s
Summary:
resolveWildcardPath is pretty hot, mostly because of the I/O it does. This
patch makes it so that in the nominal case we don't need to do any filesystem
access. That is, for every path segment, if we have no pattern matching
symbols, and we're not the final entry, then we know must be supposed to be a
dir. This avoids calling isDir and readDir for the vast majority of path
segments, including the cgroup root, which also reduces time in the kernel
doing VFS/kernfs work.

The more cgroups tracked, and the more rules loaded, the more CPU is saved. In
my tests with the example desktop ruleset and a moderately sized cgroup
hierarchy, it's about ~2%. It may be more on servers with lots of rules or lots
of cgroups.

Reviewed By: lnyng

Differential Revision: D18785019

fbshipit-source-id: 95d00a289ddd0dfd7ca6a755b2607bdf5174bf6c