Add nr_dying_descendants plugin
Current kernels have an issue where if there are too many "dying
cgroups" (ie. cgroups have not yet been reclaimed), the kernel can waste
a lot of time iterating through dying cgroups looking for pages to
reclaim. This wasted time can trigger heavy memory pressure if a process
tries to allocate a lot of memory even if the system appears to have
evictable memory.

The proper fix is for the kernel to less aggressively defer reclaim. In
the meantime, we will gate triggering kills on having <= N dying
cgroups.