Senpai detect memory.high.tmp on first access
Summary:
Currently senpai uses a hard-coded path to check if memory.high.tmp is available. This is not reliable and bad for testing.

For example, that cgroup might not exist, or it has no memory controller enabled.

Assuming any cgroup run with senpai has memory controller enabled (could be a bad assumption but not fixed here), the first access to the memory.high.tmp in that cgroup tells us if it is available system-wide.

Reviewed By: hnaz

Differential Revision: D20228077

fbshipit-source-id: 742916d694cfc95424038438243d405c68227f47