plugins respect global --cgroup-fs flag
Summary:
Plugins use the cgroup fs mount point set with the --cgroup-fs command line flag, instead of per-plugin cgroup_fs arguments.

This makes it impossible to have different plugins looking at different cgroup fs-es, which we think one would only do by accident.

Introduces ConfigCompileContext to hold cgroup_fs, which is passed to plugins on init().  We don't use OomdContext because it's created and used in the run() phase of Oomd, and we need to tell plugins who their cgroup_fs is in the compilation (init()) stage.

Reviewed By: danobi

Differential Revision: D20235419

fbshipit-source-id: 65fe0af1fadd197842611b34fb406c9569ce87be