Remove AdjustCgroup plugin
Summary:
As we don't have a contract on plugin execution order between rulesets, detector groups, and dropins, allowing any plugin to modify any shared object in arbitrary place of execution will be too fragile, i.e. some plugins will see the modification and some not.

Currently only AdjustCgroup plugin uses the mutable API, and it's not used in production. Let's removed it so we can deprecate the mutable API.

Reviewed By: danobi

Differential Revision: D21289685

fbshipit-source-id: 3a871e1d8d5b07f95b61e50a2880160fec7a5f78