Fix keeping stale cgroup
Summary:
We used to use `fstat` on cgroup dir fd to check if a cgroup is gone. However that does not work. Unfortunately, even hard link count doesn't either. As a result, staled cgroups won't be removed from the OomdContext object and thus won't got killed until Oomd is restarted.

The fix is to use `cgroup.controllers` file as an indicator to whether the cgroup still exist. As a side effect, any unit test that mocks the cgroupfs and uses `CgroupContext::refresh` needs to add a dummy `cgroup.controllers` file.

Reviewed By: dschatzberg

Differential Revision: D21752326

fbshipit-source-id: b3c0d85d6a6cf5243e85bb6ffd223510ac795157