Expose cgroup fd api in CgroupContext
Summary:
Allow plugins to retrieve the cgroup fd from a CgroupContext to check if the cgroup still exists, and do other cgroup related operations that's not supported by CgroupContext but supported by Fs utils.

The `isValid` API is replaced, which is only used by OomdContext for verifying a newly created CgroupContext. It doesn't use `Fs::isCgroupValid` because that is for checking if a valid cgroup dir fd is still valid.

Reviewed By: dschatzberg

Differential Revision: D21767818

fbshipit-source-id: 66c009ab2ff373bb5ad024ad59785125183c00e8