Senpai drives lower limit based on anon usage
Summary:
We have `limit_min_bytes` for senpai to set the memory.high lower limit on cgroups. However, this does not take into account that anon usage is often unreclaimable.

This diff makes senpai base the lower limit on anon usage instead of zero, which should prevent thrashing on services that mostly uses anon memory.

Currently swap is not taken into account. This is OK because it will make senpai less aggressive, and that's the safe direction. I will add swap support after we have scenario testings for senpai with swap enabled.

Reviewed By: dschatzberg

Differential Revision: D21827490

fbshipit-source-id: 3ff73662081c725482ad7dbdcabb3ef02f84741d