Senpai backoff on both memory and I/O pressure
Summary:
Senpai is configured to back off when some memory pressure is observed. That kind of works but it's a bit fragile.

Senpai works by evicting file cache, which may later introduce I/O and memory pressure. In edge cases (e.g. refault info lost), we may see high I/O pressure with low memory pressure. In other cases when workload itself is doing a lot of I/O, it's also reasonable to pause senpai so we don't make the situation worse. In a word, we should configure senpai to monitor and control both memory and I/O pressure.

This patch makes senpai monitor kernel moving average pressure instead of total pressure stall time, because the moving average is easy to used (stateless) and already adopted in many external monitoring facilities.

The change is only applied to the poking approach (immediate backoff), which is still experimental and will eventually replace the regular approach. This saves the cost of deployment.

Reviewed By: dschatzberg

Differential Revision: D22347206

fbshipit-source-id: 475cd749b847158fbe958a66acd958bc1c82aa0f