Fix fs race condition
Summary:
Inside Fs::readFileByLine, a cgroup could be removed during a read of its control files. In that case, a partial or empty read will be returned. As a result, functions such as `Fs::readMemcurrentAt` will cause SEGV as it assumes lines must not be empty if it's not `std::nullopt`.

Fix by checking if reads are successful.

Reviewed By: dschatzberg

Differential Revision: D22457503

fbshipit-source-id: d0953bcfac743b3df278bb9abdd4b49a3173dd75