Use string instead of hash to identify drop in configs
Summary: Drop in configs are identified by the hash of their file names, mainly used for config removal. However, it's hard to avoid hash collision and dump the names of existing drop in configs. Since the benefit of using hash instead of file names is small, let's use file names as tags.

Reviewed By: dschatzberg

Differential Revision: D22566292

fbshipit-source-id: 59df12b410873e5ff3d014163cb2a4d0bdda9788