Refactor drop in code
Summary:
Currently core oomd loop is driven by the drop in event loop, i.e. an `epoll_wait` blocks oomd core until a periodic timer or drop in events fire. As a result, drop in events trigger main oomd logic.

This is suboptimal because plugins may rely on a steady interval to calculate rate of change in counters, rate limit operations, etc. It is reasonable to separate the drop in logic from the core oomd logic.

This diff refactors the drop in logic to make drop in operations asynchronous: A drop in service communicates with external entities to append drop in config operations (add/remove) to a lock protected data structure. Oomd core loop periodically applies those operations and reports to the drop in service if they are successful. A drop in service adaptor manages the lock protected data structure and provides proper API between drop in service and oomd core.

Right now we only have one drop in service implementation, which uses file system inotify. With this framework it will be much easier to add other services such as ones over unix socket or ports.

Reviewed By: dschatzberg

Differential Revision: D22566293

fbshipit-source-id: 6e8111054685cd811b86d883e3c18d8bcf9938ce