Fix senpai immediate_backoff initializeCgroup
Summary:
When initializing cgroup state in senpai, we use to change memory high to same value of memory current to speed up convergence. However, this is no longer needed for immediate_backoff, which always sets memory limit based on memory.current reading instead of internal state.

The old behavior is unnecessary and also wrong because for immediate_backoff, we passively react to pressure, i.e. only take action when pressure is low enough. As a result, the limit set during initialization may not be reset until it expires.

Reviewed By: hnaz

Differential Revision: D22956141

fbshipit-source-id: efe8fe5a3df6ff51b5830ad70b3f06f264c4b34d