async action plugins
Summary:
Any plugin in an action chain may return ASYNC_PAUSED to pause the action chain, to be resumed at the next oomd run loop. At the next run loop, the plugin that previously returned ASYNC_PAUSED will be run() again. It may continue returning ASYNC_PAUSED as long as it likes. When it eventually returns either STOP or CONTINUE, the action chain continues executing the plugins in the chain after the async one.

The action chain will not be started (from the beginning) while paused. Its ruleset's detectors will be run every main loop, but their results will be ignored as long as the action chain is paused.

Reviewed By: dschatzberg

Differential Revision: D24743930

fbshipit-source-id: a5b2974e96c0d5dfcdf5cdbc5810c7886bce2be3