prekill hooks
Summary:
Added support for prekill hooks. See docs/prekill_hooks.md for background, mechanism, and API.

No change to oomd behavior without oomd config changes. Added "prekill_hooks" field to oomd config json schema, see docs for details.

Prekill hooks are nonblocking in kill plugins. To support this, during a kill, a KillPlugin will track a set of cgroups which it has tried and failed to kill. This set of cgroup paths should typically be a small additional memory overhead, as typically oomd will succeed in killing a cgroup on the first or second try.

Hook timeouts not implemented in this commit.

Reviewed By: lnyng

Differential Revision: D25534786

fbshipit-source-id: 739efd1e378a1b65f758492765e611e639941eb2