log unsuccessful kills
Summary:
Log to see how often kills fail, and what's happening when they do. It may be
 1) killing doesn't fail often enough to worry it as intensely as the code does
 2) when it does, moving on to the next-best choice may not be helpful anyway

We're not sure if it's useful to pick the next-best cgroup to kill if our first choice wont die. The logic is complicated by recursive killing and resuming from prekill hooks, so we'd like to remove it if possible.

Reviewed By: lnyng

Differential Revision: D26612580

fbshipit-source-id: 36b1ded077b11345d59d26003ec20dced2d3f9b4