prekill hook timeouts
Summary:
In a kill cycle, there is a max amount of time we allow before we must start killing to ensure the health of the system. This is enforced at the ruleset level: the deadline is determined when the action chain fires, and will be held across kill attempts and even kill plugins.

The default timeout is 5s, but this is configurable per-ruleset.

Reviewed By: lnyng

Differential Revision: D26667796

fbshipit-source-id: 4ae84f02a4dbffece524cc2910e7b1b5f0ae5fff