log kill failures where there are no nonempty cgroups to attempt killing
Summary:
If a kill plugin is unsuccessful trying to kill a cgroup, it logs the failure via dumpKillInfo. However, kill plugins skip trying to kill empty cgroups to avoid extraneous logs/counter bumps in killstorm situations, where a kill plugin would otherwise "re-kill" an already-killed high preference cgroup on its way down the target list to an actually interesting target.

However, if the killstorm is so bad that oomd has already killed all relevant cgroups, oomd will keep trying to kill, but not log anything at all, because it's not actually failing to kill; it simply has nothing to kill.

With this diff, a kill plugin that fails to kill will always log at least one failed kill. If there are >= 1 empty cgroups considered (and skipped for being empty), oomd will log as if it had tried and failed to kill the first (empty) cgroup it considered. If there were no cgroups at all for it to consider, it oomd will log that it failed to kill cgroup at path "".

Reviewed By: dschatzberg

Differential Revision: D29508191

fbshipit-source-id: aaf4a93898b6dda9800e63c70105ff62163cd8af