Allow SwapFree plugin to check swapout rate
Summary:
There are incidents where OOMD is triggered on low swap free, and a cgroup with high memory usage but low swap usage is killed due to kill preference config. This frees up memory but does not resolve the low swap free situation. Quickly after, OOMD is triggered again and killing the cgroup with high swap usage, even if there are still plenty of free memory.

The issue here is that freeing memory does not immediately lowers swap usage (takes time to fault pages back in). A workaround is to only trigger more kills if there are still active swapout, which indicates the system is likely in a bad state.

The default behavior of SwapFree is not changed.

Reviewed By: dschatzberg

Differential Revision: D36911756

fbshipit-source-id: 6910720fa9c69a9aab835f979f67398d69f8433f