fb-oomd: use scuba-sourced test samples for cfgen where possible
Summary:
We recently implemented the option to pull test data from scuba (dataset cfgen_inputs) instead of ssh-ing to a hardcoded host name. The benefits include security (less ssh to prod) and user experience (no need to update the test manifest when hosts die; do-not-repeat-yourself: no need for the ad hoc comments how to update the host if the scuba query is descriptive enough).

Updating fb-oomd manifest with scuba-sourced hosts where possible.

Unfortunately I couldn't do it for all hosts yet, but this should still be an improvement. I hope we will have richer data in scuba later on and be able to update the rest as well

Reviewed By: chengxiong-ruan

Differential Revision: D58533150

fbshipit-source-id: 5a6804d44f9e2e9e05e12e60e4fb2f59ae346b2d