Oomd: fix pid counting
Summary:
Previously, oomd counted the number of processes that were present in a cgroup by recursively going down and aggregating cgroup.pids. However, since this required recursive travel, oomd switched over to using pids.current instead. This change caused problems on cgroups that did not have the pid controller enabled, and caused oomd to be unable to kill.

This diff fixes this behavior by using cgroup.events to determine whether a cgroup is empty (and therefore, whether to select this cgroup), and uses either pids.current to increment the number of processes killed if possible, or returns 1 as the number of processes killed.

Reviewed By: lnyng

Differential Revision: D68645177

fbshipit-source-id: 8f8b331d3880f47bafafb19fc8c1e7a5e82b4d2b