add xattr_filter and cgroup to oomd ruleset-level config
Summary:
adds the fields xattr_filter and cgroup to ruleset level.

if there is no cgroup set at the xattr-level, ruleset will "runOnce" just like before
otherwise, we run once per cgroup that is enumerated by eval'ing the wildcard and checking xattrs for each
xattr filters check for the presence of the xattr name, ignoring xattr values

it is up to the plugin implementation to ignore or use the ruleset-level cgroup config

Rulesets now contain a "running set" of Rulesets. The parent ruleset has a copy of the detectorgroups and actions which are "cloned" by re-initing from the plugin registry. Each of these rulesets are run per cgroup that matches.
If a cgroup stops being seen (or loses its xattr), then the child ruleset is reaped.

BasePlugin now holds a copy of the PluginConstructionContext and PluginArgs so that any holder of a BasePlugin can re-init.

Reviewed By: lnyng

Differential Revision: D69419410

fbshipit-source-id: d494ff493316851923304ed262b367bbd2de0b0c