Add ruleset_cgroup to plugins
Summary: Allow plugins to specify new arg ruleset_cgroup which has same format as the existing cgroup field, but is relative to the ruleset_cgroup in OomdContext. If ruleset_cgroup is not provided, the new field is ignored. Otherwise, both regular cgroups and the ones relative to ruleset_cgroup are used for detection and killing decisions.

Differential Revision: D78754305

fbshipit-source-id: 0125306825ae59bc99ff37f4b305c14ea819ea17